fix: set color values to unsigned shorts.
The reason there are no unsigned chars is, because chars represent characters, we want numbers. I truly do not care about tricks, so even though char can go from a number and back I straight up refuse to use it like that.
There is also an int_8 type but we can not assume all compilers have it. For saving ourselves a headache I have decided not to use it.