fix: Suppress the linter on the static member.
So, clang-tidy has a skill issue in recognising static members as static members, and it thinks they are variables. Not good. So the suppression will have to do for now. Not pretty, but this is an internal problem inside clang-tidy that we can not touch. The config is set to deal with  static members properly. The linter just does not see our static member variables as statics but rather as normal variables...