fix: add test basis and tweak tests a little to function more properly.
While writing the tests I found some things that had to be improved. In the process I discovered that we have to make some mayor tweaks to the GameObject class and the Scene itself. It may be wise to give GameObject instances GUIDs, because working with names isn't going to cut it as an identifier.