feat: Alter scene manager and scene.
Alter the scene so it can no longer be instantiated by any mere mortal. Only the Scene service is worthy.

Alter the scene so it has hooks for start, stop, and end.

Alter the SceneManager so it can keep track of multiple scenes. Add additional helpers to assist with the new functionality.