feat(scene): add basic + scene manager (#96)
* feat: Scene skeleton

* feat: Implement Scene skeleton.

* feat: add game loop skeleton to scene

* feat: add game loop skeleton to scene

* feat: SceneService skeleton

* feat: Basic scene tests

* fix: Give dummy scene a proper base delegating constructor

* fix: add test basis and tweak tests a little to function more properly.

While writing the tests I found some things that had to be improved. In the process I discovered that we have to make some mayor tweaks to the GameObject class and the Scene itself. It may be wise to give GameObject instances GUIDs, because working with names isn't going to cut it as an identifier.

* feat: fix bugs based on test results.

So there were (as expected) bugs in the code:

1. The parent child relationship between for gameobjects was busted, now it isn't.
2. The engine was holding a fleeing value to a string

Linter suggestions where implemented on top of this.

* feat: Alter scene manager and scene.

Alter the scene so it can no longer be instantiated by any mere mortal. Only the Scene service is worthy.

Alter the scene so it has hooks for start, stop, and end.

Alter the SceneManager so it can keep track of multiple scenes. Add additional helpers to assist with the new functionality.